backend/pyproject.toml
[project]
authors = [
    { name = "Federico Oberndorfer", email = "federico.ober@hotmail.com" },
]
name = "api"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.1",
    "python-multipart>=0.0.6",
    "pydantic[email]>=2",
    "pydantic-settings>=2",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2",
]
version = "0.1.0"

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.19.0",
    "mypy>=1.4.1",
    "sqlalchemy[mypy]",
    "types-python-jose",
    "types-passlib",
    "ruff",
]


[build-system]
requires = ["setuptools>=68"]


[tool.mypy]
files = ["src"]
strict = true
plugins = ["sqlalchemy.ext.mypy.plugin", "pydantic.mypy"]


[tool.ruff]
select = ["I", "PGH", "S", "B"]

[tool.ruff.per-file-ignores]
"src/api/endpoints/**/*" = ["B008"]
